mcp-platform/orchestrator/requirements.txt
jinja2==3.1.2
pyyaml==6.0.1
pathlib==1.0.1